Separate out into files and deal with nans
