+++ hhs_facilities/delphi_hhs_facilities/constants.py
@@ -1,4 +1,5 @@
 """Registry for geographies to process."""
+from .generate_signals import sum_cols
 
 ADMISSION_TOTAL = "total_admissions_7dav"
 
@@ -7,7 +8,7 @@ SIGNALS = [
     (ADMISSION_TOTAL,
      ["previous_day_admission_adult_covid_confirmed_7_day_sum",
       "previous_day_admission_pediatric_covid_confirmed_7_day_sum"],
-     sum,
+     sum_cols,
      -1),
 ]
